chore(build): replace Grunt with Rollup (#77)
* Rollup Build

use prebundle semver for space savings
Add export wrapper for 4.1
ensure a context is supplied to iife so "this" is not overwritten to top level undefined.

* Update spine 3.8 to version with patches

* Modify builds as noted by vendor in readme

Vendor provides instructions for using this in esm builds.

* Update prebuild modules

* Remove favicon warning

* Remove Grunt Build System

* Remove prebuilt semver iife

iife module is no longer required

* Update package.json

one less word to slowly sound out

* Make 4.0 Support a 3 Line Chore

Add alias for 4.0.  versions below this were not published to npm.  They are available as branches in the spine-core repo.

* Add 4.0 build stub

* Update rollup.config.mjs

move example to example

* Remove spine40

"spine40": "npm:@esotericsoftware/spine-core@4.0"

Will be a part of the 4.0 PR.

Make sure we are at latest PlayCanvas code standard.

* Fix lint warning PlayCanvas Eslint 1.7.0

* Remove esbuild

Second half of build system will be in another PR.